% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{check_h3_hierarchy}
\alias{check_h3_hierarchy}
\title{Check for parent-descendant overlap}
\usage{
check_h3_hierarchy(h3)
}
\arguments{
\item{h3}{Character vector of H3 indexes.}
}
\value{
A tibble describing hierarchy overlaps.
}
\description{
Detects cases where a parent H3 cell and one of its descendants are both
present in the same H3 set.
}
