% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{h3_compaction_geometry}
\alias{h3_compaction_geometry}
\title{Create H3 compaction geometry}
\usage{
h3_compaction_geometry(
  h3,
  container = NULL,
  clip = FALSE,
  output_crs = 4326,
  analysis_crs = 7899
)
}
\arguments{
\item{h3}{H3 indexes.}

\item{container}{Optional sf or sfc polygon.}

\item{clip}{Clip geometry to the container.}

\item{output_crs}{Output CRS.}

\item{analysis_crs}{Projected CRS used for clipping.}
}
\value{
An sf object.
}
\description{
Create H3 compaction geometry
}
