% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{summarise_h3_resolution}
\alias{summarise_h3_resolution}
\title{Summarise H3 resolutions}
\usage{
summarise_h3_resolution(h3)
}
\arguments{
\item{h3}{Character vector of H3 indexes.}
}
\value{
A tibble containing resolution counts and percentages.
}
\description{
Summarise H3 resolutions
}
