% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncompact.R
\name{uncompact_h3}
\alias{uncompact_h3}
\title{Uncompact H3 cells}
\usage{
uncompact_h3(h3, resolution)
}
\arguments{
\item{h3}{H3 indexes.}

\item{resolution}{Target resolution.}
}
\value{
A character vector.
}
\description{
Uncompact H3 cells
}
