% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_h3}
\alias{validate_h3}
\title{Validate H3 indexes}
\usage{
validate_h3(h3, allow_na = FALSE, allow_duplicates = FALSE)
}
\arguments{
\item{h3}{Character vector of H3 indexes.}

\item{allow_na}{Whether missing indexes are allowed.}

\item{allow_duplicates}{Whether duplicated indexes are allowed.}
}
\value{
The validated H3 indexes as a character vector.
}
\description{
Validate H3 indexes
}
